feat: use decimal instead of native float types and optimize matching order

+++ src/components/dto.rs
@@ -1,16 +1,14 @@
 use chrono::{DateTime, Utc};
+use rust_decimal::Decimal;
 use uuid::Uuid;
 
 #[derive(Debug, Clone, Copy)]
+#[allow(dead_code)]
 pub enum TimeInForce {
-    /// Remains active until cancelled or fully filled.
-    GTC,
-    /// Executes immediately; any unfilled remainder is cancelled.
-    IOC,
-    /// Must fill in full immediately or the entire order is cancelled.
-    FOK,
-    /// Expires 24 hours after submission.
-    DAY,
+    GTC, // Good Till Cancelled
+    IOC, // Immediate Or Cancel
+    FOK, // Fill Or Kill
+    DAY, // Day Order
 }
 
 #[derive(Debug, Clone, Copy)]
@@ -20,6 +18,7 @@ pub enum OrderSide {
 }
 
 #[derive(Debug, Clone, Copy)]
+#[allow(dead_code)]
 pub enum OrderStatus {
     Open,
     PartiallyFilled,
@@ -28,12 +27,14 @@ pub enum OrderStatus {
 }
 
 #[derive(Debug, Clone, Copy)]
+#[allow(dead_code)]
 pub enum OrderType {
     Limit,
     Market,
 }
 
 #[derive(Debug, Clone)]
+#[allow(dead_code)]
 pub struct Order {
     pub id: Uuid,
     pub item_id: Uuid,
@@ -41,9 +42,9 @@ pub struct Order {
     pub order_side: OrderSide,
     pub order_type: OrderType,
     pub time_in_force: TimeInForce,
-    pub price: f32,
-    pub quantity: f32,
-    pub quantity_filled: f32,
+    pub price: Decimal,
+    pub quantity: Decimal,
+    pub quantity_filled: Decimal,
     pub status: OrderStatus,
     pub created_at: DateTime<Utc>,
     pub updated_at: DateTime<Utc>,
@@ -51,22 +52,24 @@ pub struct Order {
 }
 
 #[derive(Debug, Clone)]
+#[allow(dead_code)]
 pub struct Trade {
     pub id: Uuid,
     pub buy_order_id: Uuid,
     pub sell_order_id: Uuid,
     pub item_id: Uuid,
-    pub quantity: f32,
-    pub price: f32,
+    pub quantity: Decimal,
+    pub price: Decimal,
     pub timestamp: chrono::DateTime<Utc>,
 }
 
+#[allow(dead_code)]
 pub struct CreateOrderRequest {
     pub item_id: Uuid,
     pub user_id: Uuid,
     pub order_side: OrderSide,
     pub order_type: OrderType,
-    pub price: f32,
-    pub quantity: f32,
+    pub price: Decimal,
+    pub quantity: Decimal,
     pub time_in_force: TimeInForce,
 }
